typescript and functional components in example
