--- example/App.js
@@ -1,52 +1,0 @@
-/**
- * Sample React Native App
- *
- * adapted from App.js generated by the following command:
- *
- * react-native init example
- *
- * https://github.com/facebook/react-native
- */
-
-import React, { Component } from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
-import 'react-native-gesture-handler';
-import { SpotIMEventEmitter } from '@spot.im/react-native-spotim';
-
-import MainScreen from './Screens/MainScreen.js';
-import ArticlesScreen from './Screens/ArticlesScreen.js';
-import ArticleScreen from './Screens/ArticleScreen.js';
-import LoginScreen from './Screens/LoginScreen.js';
-
-const Stack = createStackNavigator();
-export default class App extends Component {
-  constructor(props) {
-    super(props);
-  }
-
-  onTrackAnalyticsEvent(event) {
-    console.log('\n', event.type, '\n', event);
-  }
-
-  componentDidMount() {
-    // Show analytics events
-    SpotIMEventEmitter.addListener(
-      'trackAnalyticsEvent',
-      this.onTrackAnalyticsEvent,
-    );
-  }
-
-  render() {
-    return (
-      <NavigationContainer>
-        <Stack.Navigator initialRouteName="Home">
-          <Stack.Screen name="Home" component={MainScreen} />
-          <Stack.Screen name="Articles" component={ArticlesScreen} />
-          <Stack.Screen name="Article" component={ArticleScreen} />
-          <Stack.Screen name="Login" component={LoginScreen} />
-        </Stack.Navigator>
-      </NavigationContainer>
-    );
-  }
-}
