+++ example/index.ts
@@ -1,7 +1,3 @@
-/**
- * @format
- */
-
 import { AppRegistry } from 'react-native';
 import App from './App';
 import { name as appName } from './app.json';
